Always use ensureSigner for nostr signing
This fixes a problem that occurs with muting users using secret keys.
